third_party/blink/web_tests/external/wpt/fetch/private-network-access/window-open.tentative.window-expected.txt

This is a testharness.js-based test.
[PASS] local to local: no preflight required.
[PASS] local to private: no preflight required.
[PASS] local to public: no preflight required.
[FAIL] private to local: failure.
  assert_equals: expected "failure" but got "success"
[PASS] private to private: no preflight required.
[PASS] private to public: no preflight required.
[FAIL] public to local: failure.
  assert_equals: expected "failure" but got "success"
[FAIL] public to private: failure.
  assert_equals: expected "failure" but got "success"
[PASS] public to public: no preflight required.
[FAIL] treat-as-public-address to local: failure.
  assert_equals: expected "failure" but got "success"
[FAIL] treat-as-public-address to private: failure.
  assert_equals: expected "failure" but got "success"
[PASS] treat-as-public-address to public: no preflight required.
Harness: the test ran to completion.
